+++ src/assault/game/gameObjects/Resource.java
@@ -2,7 +2,6 @@
  * To change this template, choose Tools | Templates
  * and open the template in the editor.
  */
-
 package assault.game.gameObjects;
 
 import assault.game.display.GameArea;
@@ -13,9 +12,9 @@ import assault.game.loading.resourceHolders.ResourceResourceHolder;
  *
  * @author matt
  */
-public class Resource extends Selectable{
+public class Resource extends Selectable {
 
-	public Resource(GameArea g,int x,int y, ResourceResourceHolder src) throws ResourceException {
-		super(g, x, y, src, 1, null);
-	}
+    public Resource(GameArea g, double x, double y, ResourceResourceHolder src) throws ResourceException {
+        super(g, x, y, src, 1, null);
+    }
 }
